Struct rusoto_codestar::CodeStarClient

pub struct CodeStarClient { /* fields omitted */ }

A client for the CodeStar API.

Methods

impl CodeStarClient

pub fn new(region: Region) -> CodeStarClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CodeStarClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,