Trait rusoto_codestar::CodeStar

pub trait CodeStar {
    fn associate_team_member(
        &self, 
        input: AssociateTeamMemberRequest
    ) -> RusotoFuture<AssociateTeamMemberResult, AssociateTeamMemberError>;

    fn create_project(
        &self, 
        input: CreateProjectRequest
    ) -> RusotoFuture<CreateProjectResult, CreateProjectError>;

    fn create_user_profile(
        &self, 
        input: CreateUserProfileRequest
    ) -> RusotoFuture<CreateUserProfileResult, CreateUserProfileError>;

    fn delete_project(
        &self, 
        input: DeleteProjectRequest
    ) -> RusotoFuture<DeleteProjectResult, DeleteProjectError>;

    fn delete_user_profile(
        &self, 
        input: DeleteUserProfileRequest
    ) -> RusotoFuture<DeleteUserProfileResult, DeleteUserProfileError>;

    fn describe_project(
        &self, 
        input: DescribeProjectRequest
    ) -> RusotoFuture<DescribeProjectResult, DescribeProjectError>;

    fn describe_user_profile(
        &self, 
        input: DescribeUserProfileRequest
    ) -> RusotoFuture<DescribeUserProfileResult, DescribeUserProfileError>;

    fn disassociate_team_member(
        &self, 
        input: DisassociateTeamMemberRequest
    ) -> RusotoFuture<DisassociateTeamMemberResult, DisassociateTeamMemberError>;

    fn list_projects(
        &self, 
        input: ListProjectsRequest
    ) -> RusotoFuture<ListProjectsResult, ListProjectsError>;

    fn list_resources(
        &self, 
        input: ListResourcesRequest
    ) -> RusotoFuture<ListResourcesResult, ListResourcesError>;

    fn list_tags_for_project(
        &self, 
        input: ListTagsForProjectRequest
    ) -> RusotoFuture<ListTagsForProjectResult, ListTagsForProjectError>;

    fn list_team_members(
        &self, 
        input: ListTeamMembersRequest
    ) -> RusotoFuture<ListTeamMembersResult, ListTeamMembersError>;

    fn list_user_profiles(
        &self, 
        input: ListUserProfilesRequest
    ) -> RusotoFuture<ListUserProfilesResult, ListUserProfilesError>;

    fn tag_project(
        &self, 
        input: TagProjectRequest
    ) -> RusotoFuture<TagProjectResult, TagProjectError>;

    fn untag_project(
        &self, 
        input: UntagProjectRequest
    ) -> RusotoFuture<UntagProjectResult, UntagProjectError>;

    fn update_project(
        &self, 
        input: UpdateProjectRequest
    ) -> RusotoFuture<UpdateProjectResult, UpdateProjectError>;

    fn update_team_member(
        &self, 
        input: UpdateTeamMemberRequest
    ) -> RusotoFuture<UpdateTeamMemberResult, UpdateTeamMemberError>;

    fn update_user_profile(
        &self, 
        input: UpdateUserProfileRequest
    ) -> RusotoFuture<UpdateUserProfileResult, UpdateUserProfileError>;
}

Trait representing the capabilities of the CodeStar API. CodeStar clients implement this trait.

Required methods

fn associate_team_member(
&self,
input: AssociateTeamMemberRequest
) -> RusotoFuture<AssociateTeamMemberResult, AssociateTeamMemberError>

Adds an IAM user to the team for an AWS CodeStar project.

fn create_project(
&self,
input: CreateProjectRequest
) -> RusotoFuture<CreateProjectResult, CreateProjectError>

Creates a project, including project resources. This action creates a project based on a submitted project request. A set of source code files and a toolchain template file can be included with the project request. If these are not provided, an empty project is created.

fn create_user_profile(
&self,
input: CreateUserProfileRequest
) -> RusotoFuture<CreateUserProfileResult, CreateUserProfileError>

Creates a profile for a user that includes user preferences, such as the display name and email address assocciated with the user, in AWS CodeStar. The user profile is not project-specific. Information in the user profile is displayed wherever the user's information appears to other users in AWS CodeStar.

fn delete_project(
&self,
input: DeleteProjectRequest
) -> RusotoFuture<DeleteProjectResult, DeleteProjectError>

Deletes a project, including project resources. Does not delete users associated with the project, but does delete the IAM roles that allowed access to the project.

fn delete_user_profile(
&self,
input: DeleteUserProfileRequest
) -> RusotoFuture<DeleteUserProfileResult, DeleteUserProfileError>

Deletes a user profile in AWS CodeStar, including all personal preference data associated with that profile, such as display name and email address. It does not delete the history of that user, for example the history of commits made by that user.

fn describe_project(
&self,
input: DescribeProjectRequest
) -> RusotoFuture<DescribeProjectResult, DescribeProjectError>

Describes a project and its resources.

fn describe_user_profile(
&self,
input: DescribeUserProfileRequest
) -> RusotoFuture<DescribeUserProfileResult, DescribeUserProfileError>

Describes a user in AWS CodeStar and the user attributes across all projects.

fn disassociate_team_member(
&self,
input: DisassociateTeamMemberRequest
) -> RusotoFuture<DisassociateTeamMemberResult, DisassociateTeamMemberError>

Removes a user from a project. Removing a user from a project also removes the IAM policies from that user that allowed access to the project and its resources. Disassociating a team member does not remove that user's profile from AWS CodeStar. It does not remove the user from IAM.

fn list_projects(
&self,
input: ListProjectsRequest
) -> RusotoFuture<ListProjectsResult, ListProjectsError>

Lists all projects in AWS CodeStar associated with your AWS account.

fn list_resources(
&self,
input: ListResourcesRequest
) -> RusotoFuture<ListResourcesResult, ListResourcesError>

Lists resources associated with a project in AWS CodeStar.

fn list_tags_for_project(
&self,
input: ListTagsForProjectRequest
) -> RusotoFuture<ListTagsForProjectResult, ListTagsForProjectError>

Gets the tags for a project.

fn list_team_members(
&self,
input: ListTeamMembersRequest
) -> RusotoFuture<ListTeamMembersResult, ListTeamMembersError>

Lists all team members associated with a project.

fn list_user_profiles(
&self,
input: ListUserProfilesRequest
) -> RusotoFuture<ListUserProfilesResult, ListUserProfilesError>

Lists all the user profiles configured for your AWS account in AWS CodeStar.

fn tag_project(
&self,
input: TagProjectRequest
) -> RusotoFuture<TagProjectResult, TagProjectError>

Adds tags to a project.

fn untag_project(
&self,
input: UntagProjectRequest
) -> RusotoFuture<UntagProjectResult, UntagProjectError>

Removes tags from a project.

fn update_project(
&self,
input: UpdateProjectRequest
) -> RusotoFuture<UpdateProjectResult, UpdateProjectError>

Updates a project in AWS CodeStar.

fn update_team_member(
&self,
input: UpdateTeamMemberRequest
) -> RusotoFuture<UpdateTeamMemberResult, UpdateTeamMemberError>

Updates a team member's attributes in an AWS CodeStar project. For example, you can change a team member's role in the project, or change whether they have remote access to project resources.

fn update_user_profile(
&self,
input: UpdateUserProfileRequest
) -> RusotoFuture<UpdateUserProfileResult, UpdateUserProfileError>

Updates a user's profile in AWS CodeStar. The user profile is not project-specific. Information in the user profile is displayed wherever the user's information appears to other users in AWS CodeStar.

Loading content...