Crate rusoto_codestar

AWS CodeStar
This is the API reference for AWS CodeStar. This reference provides descriptions of the operations and data types for the AWS CodeStar API along with usage examples.

You can use the AWS CodeStar API to work with:

Projects and their resources, by calling the following:

DeleteProject, which deletes a project in AWS CodeStar.

DescribeProject, which lists the attributes of a project.

ListProjects, which lists all AWS CodeStar projects associated with your AWS account.

ListResources, which lists the resources associated with an AWS CodeStar project.

UpdateProject, which updates the attributes of an AWS CodeStar project.

Teams and team members, by calling the following:

AssociateTeamMember, which adds an IAM user to the team for an AWS CodeStar project.

DisassociateTeamMember, which removes an IAM user from the team for an AWS CodeStar project.

ListTeamMembers, which lists all the IAM users in the team for an AWS CodeStar project, including their roles and attributes.

Users, by calling the following:

CreateUserProfile, which creates a user profile that contains data associated with the user across all AWS CodeStar projects.

DeleteUserProfile, which deletes all user profile information across all AWS CodeStar projects.

DescribeUserProfile, which describes the profile of a user.

ListUserProfiles, which lists all AWS CodeStar user profiles.

UpdateUserProfile, which updates the profile for an AWS CodeStar user.

If you're using the service, you're probably looking for CodeStarClient and CodeStar.

Structs

AssociateTeamMemberRequest

AssociateTeamMemberResult

CodeStarClient
A client for the CodeStar API.

CreateProjectRequest

CreateProjectResult

CreateUserProfileRequest

CreateUserProfileResult

DeleteProjectRequest

DeleteProjectResult

DeleteUserProfileRequest

DeleteUserProfileResult

DescribeProjectRequest

DescribeProjectResult

DescribeUserProfileRequest

DescribeUserProfileResult

DisassociateTeamMemberRequest

DisassociateTeamMemberResult

ListProjectsRequest

ListProjectsResult

ListResourcesRequest

ListResourcesResult

ListTeamMembersRequest

ListTeamMembersResult

ListUserProfilesRequest

ListUserProfilesResult

ProjectSummary
Information about the metadata for a project.

Resource
Information about a resource for a project.

TeamMember
Information about a team member in a project.

UpdateProjectRequest

UpdateProjectResult

UpdateTeamMemberRequest

UpdateTeamMemberResult

UpdateUserProfileRequest

UpdateUserProfileResult

UserProfileSummary
Information about a user's profile in AWS CodeStar.

Enums

AssociateTeamMemberError
Errors returned by AssociateTeamMember

CreateProjectError
Errors returned by CreateProject

CreateUserProfileError
Errors returned by CreateUserProfile

DeleteProjectError
Errors returned by DeleteProject

DeleteUserProfileError
Errors returned by DeleteUserProfile

DescribeProjectError
Errors returned by DescribeProject

DescribeUserProfileError
Errors returned by DescribeUserProfile

DisassociateTeamMemberError
Errors returned by DisassociateTeamMember

ListProjectsError
Errors returned by ListProjects

ListResourcesError
Errors returned by ListResources

ListTeamMembersError
Errors returned by ListTeamMembers

ListUserProfilesError
Errors returned by ListUserProfiles

UpdateProjectError
Errors returned by UpdateProject

UpdateTeamMemberError
Errors returned by UpdateTeamMember

UpdateUserProfileError
Errors returned by UpdateUserProfile

Traits

CodeStar
Trait representing the capabilities of the CodeStar API. CodeStar clients implement this trait.